split node
