+++ src/features/structure/index.js
@@ -2,3 +2,5 @@ export {default as useActions} from './actionCreators';
 export {default as useChildrenCompletion} from './useChildrenCompletion';
 export {default as useFocusedParentId} from './useFocusedParentId';
 export {default as StructureReducer} from './reducer';
+export {default as useChildrenIdList} from './useChildrenIdList';
+export {default as useNodeData} from './useNodeData';
