+++ src/features/structure/operations/getChildrenIdList.js
@@ -1,0 +1,6 @@
+import {path} from 'ramda';
+
+const getChildrenIdList = (structure, nodeId) =>
+  path(['nodes', nodeId, 'children'], structure);
+
+export default getChildrenIdList;
