+++ src/features/structure/operations/index.js
@@ -3,3 +3,4 @@ export {default as getDescendantsIdList} from './getDescendantsIdList';
 export {default as editChildren} from './editChildren';
 export {default as deleteNode} from './deleteNode';
 export {default as getChildrenCompletion} from './getChildrenCompletion';
+export {default as getChildrenIdList} from './getChildrenIdList';
