+++ src/features/structure/useNodeData.js
@@ -1,0 +1,5 @@
+import {useSelector} from 'react-redux';
+
+const useNodeData = id => useSelector(state => state.structure.nodes[id]);
+
+export default useNodeData;
